Organize files, start aromaticity detection
